+++ tests/index.js
@@ -29,7 +29,7 @@ describe('half circle progress', () => {
     const div = document.createElement(div);
     document.body.appendChild(div);
     const circle = ReactDOM.render(
-      <Circle percent="30" strokeWidth="1" gapWidth={70} gapPosition="bottom"/>,
+      <Circle percent="30" strokeWidth="1" gapDegree={70} gapPosition="bottom"/>,
       div);
     expect(circle.props.percent).to.be('30');
     ReactDOM.unmountComponentAtNode(div);
@@ -41,7 +41,7 @@ describe('half circle progress', () => {
     const div = document.createElement(div);
     document.body.appendChild(div);
     const circle = ReactDOM.render(
-      <Circle percent="30" strokeWidth="1" gapWidth={70} gapPosition="left"/>,
+      <Circle percent="30" strokeWidth="1" gapDegree={70} gapPosition="left"/>,
       div);
     expect(circle.props.percent).to.be('30');
     ReactDOM.unmountComponentAtNode(div);
@@ -53,7 +53,7 @@ describe('half circle progress', () => {
     const div = document.createElement(div);
     document.body.appendChild(div);
     const circle = ReactDOM.render(
-      <Circle percent="30" strokeWidth="1" gapWidth={70} gapPosition="right"/>,
+      <Circle percent="30" strokeWidth="1" gapDegree={70} gapPosition="right"/>,
       div);
     expect(circle.props.percent).to.be('30');
     ReactDOM.unmountComponentAtNode(div);